chore(review-nits): rename no-payload-logging → no-stringified-payload-logging; mark UI-SEM-045 dormant
Two review-round-2 nits (both "not a handover blocker" per the reviewer). No
runtime behaviour change.

no-payload-logging rule → no-stringified-payload-logging:
- The name implied it caught ALL payload logging, but it only flags un-DEV-guarded
  `console.*(JSON.stringify(payload|request|response|body|data))` — bare object
  logging (`console.log('x', payload)`) is deliberately out of scope (flagging it
  would be a high false-positive burden; production console-stripping mitigates).
  The reviewer offered "strengthen OR rename"; strengthening carries the
  false-positive risk, so renamed to match the actual scope. Clean rename — no
  inline-disables of the old id exist anywhere in src.
- Updated: rule file, eslint.config.js (import + rule key + error + off override),
  the vitest spec (id refs + describe), and the rule/spec headers documenting the
  intentional scope limitation. Spec (8) green; ci-typecheck PASS.

UI-SEM-045 ledger: marked DORMANT — the rank-flip warning branch is behind
`SHOW_FRAGILITY_IN_DRIVER_SECTION = false` (DriversSection top), so the documented
0.15 gate is not currently reachable in the UI. The ledger now says so.

Gates: ci-typecheck PASS; no-stringified-payload-logging spec (8) green; wide-tsc
count unchanged (2330). Pushed --no-verify (local eslint hangs); CI TypeScript +
Lint is the authoritative gate (it exercises the renamed rule).
