+++ CLAUDE.md
@@ -208,7 +208,7 @@ pnpm run build                      # production build (separate concern)
 | UI-SEM-042 | _`HeroSection.tsx` removed (288a2b74)_ | Fragility ratio threshold (>0.7) for trust reason | **Retired with the component** — analysis-hero consumes a producer `trustLine` (null on the live path, `buildHeroModel.ts`), not a threshold-derived reason. ID retained for history |
 | UI-SEM-043 | _`HeroSection.tsx` removed (288a2b74)_ | Evidence quality threshold (<0.5) for trust reason | **Retired with the component** — analysis-hero consumes a producer `trustLine` (null on the live path), not a threshold-derived reason. ID retained for history |
 | UI-SEM-044 | `src/lib/stability.ts` (`getStabilityClassification` → `borderClass`) | Border colour classification from stability | Keep — display formatting. **CONSOLIDATED (ledger corrected 2026-07-14 — was mislabelled "retired")**: consolidated (with 005/041) into `getStabilityClassification`; the `borderClass` it returns is LIVE, consumed by GoalNode's border override |
-| UI-SEM-045 | `src/components/results/DriversSection.tsx` (rank-flip warning) | Rank-flip warning visibility gate (`rankFlipRate >= 0.15`) | Remove when PLoT provides visibility gate. (Ledger corrected 2026-07-14 — threshold was 0.3, is now 0.15; source re-tagged) |
+| UI-SEM-045 | `src/components/results/DriversSection.tsx` (rank-flip warning) | Rank-flip warning visibility gate (`rankFlipRate >= 0.15`) | **DORMANT** — the branch is behind `SHOW_FRAGILITY_IN_DRIVER_SECTION = false` (DriversSection top), so the 0.15 gate is NOT currently reachable in the UI. Remove when PLoT provides a visibility gate. (Ledger corrected 2026-07-14 — threshold was 0.3, is now 0.15; source re-tagged; dormancy noted per review round 2) |
 | UI-SEM-046 | `src/components/results/DriversSection.tsx` (`tooltipElasticityCopy`) | Elasticity display scaling (×10, floor 1) | Remove when PLoT provides shift percentage. (Source re-tagged 2026-07-14) |
 | UI-SEM-047 | `src/components/results/DriversSection.tsx:356` | Confidence clamped to [0, 1] | Keep — normalisation |
 | UI-SEM-049 | `src/canvas/components/ModelTabBody.tsx` | VOI fallback: value_of_information * 100 as pp when evpi_percentage_points absent | Remove when PLoT guarantees evpi_percentage_points |
